[EXE] Automap: Don't close automap when pressing spacebar with other windows open (#6713)
* GameMenu: Add a flag to notify that game menu is on/off

This patch adds a simple flag so other source files can see if game menu
is currently off or on.

* Automap: Don't close automap when pressing space with other windows open

Currently if we open a lot of windows and have automap enabled, and we
press space - all of the windows along with automap get closed. This
patch removes that, and allows to close automap with spacebar, but only
if every window is closed.