[DATA] Fix not being able to equip two-handed items in a certain case
Fix a bug that caused two-handed items not to able to be equipped
if there was an item in the right hand, but not in the left one
and also there wasn't enough space to place the right item TWO times.
This was unintended, obviously, the right item should only be checked for once.

Co-authored-by: Tully <166401925+Tully-B@users.noreply.github.com>