[EXE] CMake: FetchContent `EXCLUDE_FROM_ALL` fixes
On CMake v3.28+, `EXCLUDE_FROM_ALL` is supported natively as an argument
to `FetchContent_Declare`.

On CMake v3.30+, `FetchContent_Populate`, which we use to polyfill
`EXCLUDE_FROM_ALL` support for older versions of CMake, is deprecated
and triggers a noisy warning.

Avoids the warning by using the native `EXCLUDE_FROM_ALL` support on
CMake v3.28+.